move file_watcher back to the cli crate
