+++ cli/util/file_watcher.rs
@@ -14,9 +14,6 @@ use deno_core::error::CoreError;
 use deno_core::futures::Future;
 use deno_core::futures::FutureExt;
 use deno_core::parking_lot::Mutex;
-use deno_lib::util::file_watcher::WatcherCommunicator;
-use deno_lib::util::file_watcher::WatcherCommunicatorOptions;
-use deno_lib::util::file_watcher::WatcherRestartMode;
 use deno_runtime::fmt_errors::format_js_error;
 use log::info;
 use notify::event::Event as NotifyEvent;
@@ -28,6 +25,7 @@ use notify::Watcher;
 use tokio::select;
 use tokio::sync::broadcast::error::RecvError;
 use tokio::sync::mpsc;
+use tokio::sync::mpsc::error::SendError;
 use tokio::sync::mpsc::UnboundedReceiver;
 use tokio::time::sleep;
 
@@ -144,6 +142,104 @@ fn create_print_after_restart_fn(clear_screen: bool) -> impl Fn() {
   }
 }
 
+#[derive(Debug)]
+pub struct WatcherCommunicatorOptions {
+  /// Send a list of paths that should be watched for changes.
+  pub paths_to_watch_tx: tokio::sync::mpsc::UnboundedSender<Vec<PathBuf>>,
+  /// Listen for a list of paths that were changed.
+  pub changed_paths_rx: tokio::sync::broadcast::Receiver<Option<Vec<PathBuf>>>,
+  pub changed_paths_tx: tokio::sync::broadcast::Sender<Option<Vec<PathBuf>>>,
+  /// Send a message to force a restart.
+  pub restart_tx: tokio::sync::mpsc::UnboundedSender<()>,
+  pub restart_mode: WatcherRestartMode,
+  pub banner: String,
+}
+
+/// An interface to interact with Deno's CLI file watcher.
+#[derive(Debug)]
+pub struct WatcherCommunicator {
+  /// Send a list of paths that should be watched for changes.
+  paths_to_watch_tx: tokio::sync::mpsc::UnboundedSender<Vec<PathBuf>>,
+  /// Listen for a list of paths that were changed.
+  changed_paths_rx: tokio::sync::broadcast::Receiver<Option<Vec<PathBuf>>>,
+  changed_paths_tx: tokio::sync::broadcast::Sender<Option<Vec<PathBuf>>>,
+  /// Send a message to force a restart.
+  restart_tx: tokio::sync::mpsc::UnboundedSender<()>,
+  restart_mode: Mutex<WatcherRestartMode>,
+  banner: String,
+}
+
+impl WatcherCommunicator {
+  pub fn new(options: WatcherCommunicatorOptions) -> Self {
+    Self {
+      paths_to_watch_tx: options.paths_to_watch_tx,
+      changed_paths_rx: options.changed_paths_rx,
+      changed_paths_tx: options.changed_paths_tx,
+      restart_tx: options.restart_tx,
+      restart_mode: Mutex::new(options.restart_mode),
+      banner: options.banner,
+    }
+  }
+
+  pub fn watch_paths(
+    &self,
+    paths: Vec<PathBuf>,
+  ) -> Result<(), SendError<Vec<PathBuf>>> {
+    if paths.is_empty() {
+      return Ok(());
+    }
+    self.paths_to_watch_tx.send(paths)
+  }
+
+  pub fn force_restart(&self) -> Result<(), SendError<()>> {
+    // Change back to automatic mode, so that HMR can set up watching
+    // from scratch.
+    *self.restart_mode.lock() = WatcherRestartMode::Automatic;
+    self.restart_tx.send(())
+  }
+
+  pub async fn watch_for_changed_paths(
+    &self,
+  ) -> Result<Option<Vec<PathBuf>>, RecvError> {
+    let mut rx = self.changed_paths_rx.resubscribe();
+    rx.recv().await
+  }
+
+  pub fn change_restart_mode(&self, restart_mode: WatcherRestartMode) {
+    *self.restart_mode.lock() = restart_mode;
+  }
+
+  pub fn send(
+    &self,
+    paths: Option<Vec<PathBuf>>,
+  ) -> Result<(), SendError<Option<Vec<PathBuf>>>> {
+    match *self.restart_mode.lock() {
+      WatcherRestartMode::Automatic => {
+        self.restart_tx.send(()).map_err(|_| SendError(None))
+      }
+      WatcherRestartMode::Manual => self
+        .changed_paths_tx
+        .send(paths)
+        .map(|_| ())
+        .map_err(|e| SendError(e.0)),
+    }
+  }
+
+  pub fn print(&self, msg: String) {
+    log::info!("{} {}", self.banner, colors::gray(msg));
+  }
+
+  pub fn show_path_changed(&self, changed_paths: Option<Vec<PathBuf>>) {
+    if let Some(paths) = changed_paths {
+      if !paths.is_empty() {
+        self.print(format!("Restarting! File change detected: {:?}", paths[0]))
+      } else {
+        self.print("Restarting! File change detected.".to_string())
+      }
+    }
+  }
+}
+
 /// Creates a file watcher.
 ///
 /// - `operation` is the actual operation we want to run every time the watcher detects file
@@ -171,6 +267,16 @@ where
   .boxed_local();
 
   fut.await
+}
+
+#[derive(Clone, Copy, Debug)]
+pub enum WatcherRestartMode {
+  /// When a file path changes the process is restarted.
+  Automatic,
+
+  /// When a file path changes the caller will trigger a restart, using
+  /// `WatcherInterface.restart_tx`.
+  Manual,
 }
 
 /// Creates a file watcher.
