--- lib/util/mod.rs
@@ -1,3 +1,0 @@
-// Copyright 2018-2025 the Deno authors. MIT license.
-
-pub mod file_watcher;
